fix mistake in copying files
